% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rust_interface.R
\name{get_groups_in_codes_file}
\alias{get_groups_in_codes_file}
\title{Get the list of valid group names defined in a codes file}
\usage{
get_groups_in_codes_file(codes_file_path)
}
\description{
Get the list of valid group names defined in a codes file
}
