% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_get_groups_in_codes_file}
\alias{rust_get_groups_in_codes_file}
\title{Get the code groups defined in a codes file}
\usage{
rust_get_groups_in_codes_file(codes_file_path)
}
\description{
Returns a character vector of group names defined in
the codes file. This can be used as the basis for fetching
all the code groups using rust_get_codes_in_group.
}
